causal_rate_attribution/requirements.txt
dowhy>=0.11
econml>=0.15
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
statsmodels>=0.14
lightgbm>=4.0
plotly>=5.18
pyyaml>=6.0
python-dotenv>=1.0
joblib>=1.3
scipy>=1.11
